Refactor to move file
